EAA update controllers WorkOrder

+++ src/controllers/workOrder.controller.js
@@ -8,9 +8,31 @@ const jwt = require('jsonwebtoken');
 
 // Get all work orders
 const getAllWorkOrders = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor', 'technician'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor o técnico'
+        });
+    }
+
     try {
-        const workOrders = await WorkOrder.find();
-        res.status(200).json(workOrders);
+        if (userDataToken.userRole === 'supervisor') {
+            const workOrders = await WorkOrder.find({ workOrderSupervisor: userDataToken._id });
+            return res.status(200).json(
+                { ok: true, message: 'Ordenes de trabajo encontradas', data: workOrders }
+            );
+        }
+        if (userDataToken.userRole === 'technician') {
+            const workOrders = await WorkOrder.find({ workOrderAssignedTechnician: userDataToken._id });
+            return res.status(200).json(
+                { ok: true, message: 'Ordenes de trabajo encontradas', data: workOrders }
+            );
+        }
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -18,6 +40,17 @@ const getAllWorkOrders = async (req, res) => {
 
 // Get a single work order by ID
 const getWorkOrderById = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor', 'technician'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor o técnico'
+        });
+    }
     try {
         const workOrder = await WorkOrder.findById(req.params.id);
         if (!workOrder) {
@@ -31,6 +64,17 @@ const getWorkOrderById = async (req, res) => {
 
 // Create a new work order
 const createWorkOrder = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor'
+        });
+    }
     const workOrder = req.body;
     console.log('createWorkOrder ', req.body, workOrder);
     try {
@@ -79,7 +123,7 @@ const createWorkOrder = async (req, res) => {
 // Update an existing work order
 const updateWorkOrder = async (req, res) => {
     try {
-        const updatedWorkOrder = await WorkOrder.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        const updatedWorkOrder = await WorkOrder.findByIdAndUpdate(req.params.id, req.body);
         if (!updatedWorkOrder) {
             return res.status(404).json({ message: 'WorkOrder not found' });
         }
@@ -91,6 +135,17 @@ const updateWorkOrder = async (req, res) => {
 
 // Delete a work order
 const deleteWorkOrder = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor'
+        });
+    }
     try {
         const deletedWorkOrder = await WorkOrder.findByIdAndDelete(req.params.id);
         if (!deletedWorkOrder) {
@@ -104,6 +159,17 @@ const deleteWorkOrder = async (req, res) => {
 
 // Get all work orders by client ID
 const getAllWorkOrdersByClient = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor'
+        });
+    }
     try {
         const workOrders = await WorkOrder.find({ clientId: req.params.clientId });
         res.status(200).json(workOrders);
@@ -114,6 +180,23 @@ const getAllWorkOrdersByClient = async (req, res) => {
 
 // Get all work orders by technician ID
 const getAllWorkOrdersByTechnician = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor', 'technician'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor o técnico'
+        });
+    }
+    if (userDataToken.userRole === 'technician' && userDataToken._id !== req.params.technicianId) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El técnico no tiene asignada esta orden de trabajo'
+        });
+    }
     try {
         const workOrders = await WorkOrder.find({ workOrderAssignedTechnician: req.params.technicianId });
         res.status(200).json(workOrders);
@@ -122,13 +205,29 @@ const getAllWorkOrdersByTechnician = async (req, res) => {
     }
 };
 
-
 // Get work order with tasks and task evidences
 const getReportWorkOrder = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor', 'technician'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor o técnico'
+        });
+    }
     try {
         const workOrder = await WorkOrder.findById(req.params.id);
         if (!workOrder) {
             return res.status(404).json({ message: 'WorkOrder not found' });
+        }
+        if (userDataToken.userRole === 'technician' && userDataToken._id !== workOrder.workOrderAssignedTechnician) {
+            return res.status(400).json({
+                ok: false,
+                error: 'El técnico no tiene asignada esta orden de trabajo'
+            });
         }
 
         const tasks = await WorkOrderTask.find({ workOrderId: req.params.id });
@@ -142,7 +241,6 @@ const getReportWorkOrder = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 };
-
 
 const registerAuditLog = async (req, action, documentId, changes) => {
     const token = req.header('Authorization')?.split(' ')[1];
@@ -165,6 +263,159 @@ const registerAuditLog = async (req, action, documentId, changes) => {
     await AuditLogController.createAuditLog(auditLogData);
 };
 
+
+// Update a workOrder by id
+const updateWorkOrderStatus = async (req, res) => {
+    const { id } = req.params;
+    const { workOrderStatus, workOrderReasonRejection } = req.body;
+    let hasChanges = false;
+    try {
+        const token = req.header('Authorization')?.split(' ')[1];
+        const secret = process.env.SECRET_KEY;
+        const decoded = jwt.verify(token, secret);
+        const userDataToken = await User.findById(decoded.workOrderData);
+        const validRoles = ['supervisor', 'technician'];
+        if (!validRoles.includes(userDataToken.userRole)) {
+            return res.status(400).json({
+                ok: false,
+                error: 'El rol debe ser supervisor o técnico'
+            });
+        }
+
+        const updateDataById = {};
+        const changes = {};
+        const originalData = await WorkOrder.findById(id).lean();
+        if (!originalData)
+            return res.status(400).json({
+                ok: false,
+                message: 'No se encontró ningún Orden de trabajo con el id proporcionado'
+            })
+        if (originalData) {
+            if (workOrderStatus) {
+                if (userDataToken.userRole === 'technician') {
+                    if (originalData.workOrderAssignedTechnician !== userDataToken._id)
+                        return res.status(400).json({
+                            ok: false,
+                            error: 'El técnico no tiene asignada esta orden de trabajo'
+                        });
+                    const validStatus = ['Assigned', 'In Progress', 'Under Review'];
+                    if (!validStatus.includes(workOrderStatus)) {
+                        return res.status(400).json({
+                            ok: false,
+                            error: 'El estado debe ser Asignado, En progreso o En revisión'
+                        });
+                    }
+                    updateDataById.workOrderStatus = workOrderStatus;
+                    if (originalData.workOrderReasonRejection)
+                        updateDataById.workOrderReasonRejection = null;
+                }
+                if (userDataToken.userRole === 'supervisor') {
+                    const validStatus = ['Unassigned', 'Assigned', 'Approved'];
+                    if (!validStatus.includes(workOrderStatus)) {
+                        return res.status(400).json({
+                            ok: false,
+                            error: 'El estado debe ser Sin asignar, Asignado o Aprobado'
+                        });
+                    }
+                    updateDataById.workOrderStatus = workOrderStatus;
+                }
+            }
+            if (workOrderReasonRejection && userDataToken.userRole === 'supervisor')
+                updateDataById.workOrderReasonRejection = workOrderReasonRejection;
+            if (userDataToken.userRole === 'supervisor' && workOrderStatus === 'Approved' && originalData.workOrderReasonRejection) {
+                updateDataById.workOrderReasonRejection = null;
+                // Identify changes
+                for (let key in updateDataById) {
+                    if (originalData[key] !== updateDataById[key]) {
+                        hasChanges = true;
+                        changes[key] = { old: originalData[key], new: updateDataById[key] }
+                        console.log('changes', changes);
+                    }
+                }
+            }
+
+            if (!hasChanges)
+                return res.status(400).json({
+                    ok: false,
+                    message: 'No se detectaron cambios, no se puede actualizar el Orden de trabajo'
+                })
+            const workOrder = await WorkOrder.findByIdAndUpdate(id, updateDataById)
+            if (!workOrder)
+                return res.status(400).json({
+                    ok: false,
+                    message: 'No se puede actualizar el Orden de trabajo, no encontrado o no se detectaron cambios'
+                })
+            // Register in audit_logs (req, action, documentId, changes) 
+            await registerAuditLog(req, 'UPDATE', id, { updateRecord: changes });
+
+            return res.status(200).json({
+                ok: true,
+                message: 'Estado Orden de trabajo actualizado',
+                data: workOrder
+            })
+        }
+    }
+    catch (error) {
+        console.log(error)
+        return res.status(500).json({
+            ok: false,
+            message: 'No se puede actualizar el Orden de trabajo, por favor contacte al soporte',
+            data: error
+        })
+    }
+}
+
+// Get all work orders under review with rejection reason for a technician
+const getAllWorkOrdersWithRejection = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['technician'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser técnico'
+        });
+    }
+    try {
+        const workOrders = await WorkOrder.find({
+            workOrderAssignedTechnician: userDataToken._id,
+            workOrderReasonRejection: { $exists: true, $ne: null }
+        });
+        res.status(200).json(
+            { ok: true, message: 'Reporte de Ordenes de trabajo rechazados', data: workOrders });
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
+};
+
+// Get all pending work orders to be approved by supervisor
+const getAllWorkOrdersPendingToApprove = async (req, res) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    const secret = process.env.SECRET_KEY;
+    const decoded = jwt.verify(token, secret);
+    const userDataToken = await User.findById(decoded.workOrderData);
+    const validRoles = ['supervisor'];
+    if (!validRoles.includes(userDataToken.userRole)) {
+        return res.status(400).json({
+            ok: false,
+            error: 'El rol debe ser supervisor'
+        });
+    }
+    try {
+        const workOrders = await WorkOrder.find({
+            workOrderSupervisor: userDataToken._id,
+            workOrderStatus: 'Under Review',
+            workOrderReasonRejection: null
+        });
+        res.status(200).json(
+            { ok: true, message: 'Ordenes de trabajo pendientes de aprobación encontradas', data: workOrders });
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
+};
+
 module.exports = {
     getAllWorkOrders,
     getWorkOrderById,
@@ -173,7 +424,8 @@ module.exports = {
     deleteWorkOrder,
     getAllWorkOrdersByClient,
     getAllWorkOrdersByTechnician,
-    getReportWorkOrder
+    getReportWorkOrder,
+    updateWorkOrderStatus,
+    getAllWorkOrdersWithRejection,
+    getAllWorkOrdersPendingToApprove
 };
-
-
